fixup "dde_linux: update usb_net driver to Linux 6.1.20" (prevent format warning)
Unfortunately, the Linux kernel decides to typedef __kernel_ssize_t as
'int' on platforms BITS_PER_LONG != 64 and long otherwise...

issue #4958

+++ repos/dde_linux/src/drivers/usb_net/lx_user.c
@@ -319,7 +319,7 @@ int lx_wdm_write(void *args)
 		length = wdm_driver->fops->write(&wdm_file, wdm_data->buffer,
 		                                 *wdm_data->data_avail, NULL);
 		if (length < 0) {
-			printk("WDM write error: %ld", length);
+			printk("WDM write error: %ld", (long)length);
 		}
 
 		lx_wdm_schedule_read(wdm_data->handle);
